+++ lib/src/db/lib_prod_archives.dart
@@ -5,6 +5,7 @@ import 'dart:io';
 import 'package:elec_server/src/db/cme/cme_energy_settlements.dart';
 import 'package:elec_server/src/db/config.dart';
 import 'package:elec_server/src/db/polygraph/polygraph_archive.dart';
+import 'package:elec_server/src/db/utilities/eversource/supplier_backlog_rates.dart';
 import 'package:elec_server/src/db/utilities/retail_suppliers_offers_archive.dart';
 
 CmeSettlementsEnergyArchive getCmeEnergySettlementsArchive() {
@@ -16,6 +17,17 @@ CmeSettlementsEnergyArchive getCmeEnergySettlementsArchive() {
     Directory(dir).createSync(recursive: true);
   }
   return CmeSettlementsEnergyArchive(dbConfig: dbConfig, dir: dir);
+}
+
+CtSupplierBacklogRatesArchive getCtSupplierBacklogRatesArchive() {
+  var dbConfig = ComponentConfig(
+      host: '127.0.0.1', dbName: 'retail_suppliers',
+      collectionName: 'ct_backlog_rates');
+  var archive = CtSupplierBacklogRatesArchive(dbConfig: dbConfig);
+  if (!Directory(archive.dir).existsSync()) {
+    Directory(archive.dir).createSync(recursive: true);
+  }
+  return archive;
 }
 
 PolygraphArchive getPolygraphArchive() {